backend/requirements.txt
# FastAPI & server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
alembic==1.14.1
asyncpg==0.30.0
aiosqlite==0.20.0

# Async
celery[redis]==5.4.0
redis==5.2.1

# Proxmox
proxmoxer==2.1.0
requests==2.32.3

# Ansible (subprocess runner)
ansible-runner==2.4.0

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# WebSocket
websockets==14.1

# Utils
pyyaml==6.0.2
jinja2==3.1.5
python-dotenv==1.0.1
httpx==0.28.1
loguru==0.7.3
